fix(@pixcelmon/pikachu): packaging
