+++ packages/pikachu/query-tabs/public_api.ts
@@ -1,2 +1,3 @@
-export { QueryTabsModule, QueryTab } from './query-tabs.module';
+export { QueryTabsModule } from './query-tabs.module';
 export { QueryTabsComponent } from './query-tabs.component';
+export * from './query-tabs-interface';
