+++ packages/pikachu/query-tabs/query-tabs.module.ts
@@ -10,13 +10,3 @@ import { QueryTabsComponent } from './query-tabs.component';
   exports: [CommonModule, TranslatePipeModule, FilterPipeModule, NgZorroAntdModule, QueryTabsComponent],
 })
 export class QueryTabsModule {}
-
-export interface QueryTab {
-  title: string;
-  field: string;
-  showFilter?: boolean;
-  searchValue?: any; // 搜索值
-  defaultValue?: any; // 默认值
-  displayValue?: any; // 展示值
-  lexicon?: { value: any; label: string }[]; // 词典
-}
