+++ packages/pikachu/upload/public_api.ts
@@ -1,1 +1,2 @@
 export { UploadModule } from './upload.module';
+export * from './upload-interface';
